refactor: rename transactions.rs to tx.rs
