+++ bitcoin/src/types/mod.rs
@@ -4,7 +4,7 @@
 
 pub mod legacy;
 pub mod script;
-pub mod transactions;
+pub mod tx;
 pub mod txin;
 pub mod txout;
 pub mod utxo;
@@ -12,7 +12,7 @@ pub mod witness;
 
 pub use legacy::*;
 pub use script::*;
-pub use transactions::*;
+pub use tx::*;
 pub use txin::*;
 pub use txout::*;
 pub use utxo::*;
